Add promise/future to InputEventSender

+++ LandingTrainer/SimVarSetter.h
@@ -10,7 +10,7 @@
 class SimVarSetter
 {
 public:
-	static void setSimVar(HANDLE sim, std::string name, std::string unitType, SIMCONNECT_DATATYPE dataType, std::any data, size_t dataSize, SIMCONNECT_OBJECT_ID objectID = SIMCONNECT_OBJECT_ID_USER)
+	static void setSimVar(HANDLE sim, std::string name, std::string unitType, SIMCONNECT_DATATYPE dataType, std::any data, DWORD dataSize, SIMCONNECT_OBJECT_ID objectID = SIMCONNECT_OBJECT_ID_USER)
 	{
 		DWORD defineID = IDCounter::getID();
 		SimConnect_AddToDataDefinition(sim, defineID, name.c_str(), unitType.c_str(), dataType);
